services/devicemanagement/requirements.txt
numpy~=1.19.2
redis==3.4.1
celery==4.4.2
pandas>=1.2.5
pika==1.2.0
gevent==21.1.2
pytest~=6.2.5
python-dotenv~=0.20.0
